requirements.txt
flask=1.1.1
wordcloud=1.5.0
matplotlib==2.2.2
Pillow=6.1.0
numpy=1.16.4